remove old type system
but add information if key is binary or not
remove typing information
remove type type_t
remove type from key struct
dont parse type information in xml file
keyNew now supports KEY_BINARY

+++ src/libelektra/keyset.c
@@ -598,8 +598,8 @@ ssize_t ksAppendKey(KeySet *ks, Key *toAppend)
 ssize_t ksAppend(KeySet *ks, const KeySet *toAppend)
 {
 	size_t oldSize = 0;
-	int i = 0;
-	int toAlloc = 0;
+	size_t i = 0;
+	size_t toAlloc = 0;
 
 	if (!ks) return -1;
 	if (!toAppend) return -1;
